NIFI-9430: Incorporated review comments
