+++ c2/c2-protocol/pom.xml
@@ -14,8 +14,8 @@ distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
- -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+-->
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="https://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="https://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
@@ -28,10 +28,7 @@ limitations under the License.
     <packaging>pom</packaging>
 
     <modules>
+        <module>c2-protocol-component-api</module>
         <module>c2-protocol-api</module>
-        <!--<module>efm-c2-protocol-framework</module>
-        <module>efm-c2-protocol-framework-api</module>
-        <module>efm-c2-protocol-web</module>
-        <module>efm-c2-protocol-server-api</module> -->
     </modules>
 </project>
